fix remaining build warnings better libad-test usage example replace tabs with spaces update year to 2021

+++ ad-test.c
@@ -3,7 +3,7 @@
  *
  * Example: ./ad-test-dbg 6 2 1 >/tmp/ad-test.txt; diff --context=1 /tmp/ad-test.txt ad-test.txt
  *
- * (c) 2018-2020 m4r35n357@gmail.com (Ian Smith), for licencing see the LICENCE file
+ * (c) 2018-2021 m4r35n357@gmail.com (Ian Smith), for licencing see the LICENCE file
  */
 
 
@@ -65,13 +65,13 @@ void ode (series x, series y, series z, components *c, void *params, int k) {
 }
 
 int main (int argc, char **argv) {
-    long n;
+    int n;
     mpfr_t x, y, _;
 
     assert(argc == 4);
     mpfr_set_default_prec(113);
     ad_test_tempvars();
-    n = strtol(argv[1], NULL, BASE) + 1;
+    n = (int)strtol(argv[1], NULL, BASE) + 1;
     assert(n > 1);
     mpfr_init_set_str(x, argv[2], BASE, RND);
     mpfr_init_set_str(y, argv[3], BASE, RND);
